[Move #292] 방 만들기 모달창 main 폴더로 이동
